pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-multi-agent-business-intelligence-system"
version = "2.0.0"
description = "Autonomous Multi-Agent Business Intelligence System - Multi-agent SQL generation, analytics, monitoring, and reporting"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Sabarish-29" }]
keywords = ["text-to-sql", "fastapi", "crewai", "rag", "analytics"]

# Core runtime dependencies (production)
dependencies = [
  # Web/API
  "fastapi>=0.109.0,<1.0.0",
  "uvicorn[standard]>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
  "pydantic>=2.6.0,<3.0.0",
  "pydantic-settings>=2.1.0,<3.0.0",
  "httpx>=0.26.0,<1.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "requests>=2.31.0,<3.0.0",
  "tenacity>=8.2.0,<9.0.0",
  "sqlalchemy>=2.0.25,<3.0.0",
  "websockets>=12.0,<13.0",

  # CrewAI + LLM clients
  "crewai>=1.8.0,<2.0.0",
  "crewai-tools>=0.76.0,<1.0.0",
  "openai>=1.83.0,<2.0.0",
  "anthropic>=0.25.0,<1.0.0",
  "groq>=0.30.0,<1.0.0",

  # LangChain 0.3.x ecosystem alignment
  "langchain>=0.3.0,<0.4.0",
  "langchain-core>=0.3.0,<0.4.0",
  "langchain-community>=0.3.0,<0.4.0",
  "langchain-openai>=0.3.0,<0.4.0",
  "langchain-groq>=0.3.0,<0.4.0",
  "langchain-anthropic>=0.3.0,<0.4.0",

  # Vector DB (CrewAI 1.x requires chromadb>=1.1)
  "chromadb>=1.1.0,<1.2.0",

  # Sandbox
  "docker>=7.1.0,<8.0.0",
  "RestrictedPython>=6.0,<7.0",

  # Monitoring / scheduling
  "apscheduler>=3.10.0,<4.0.0",
  "tavily-python>=0.3.0,<1.0.0",

  # UI
  "streamlit>=1.31.0,<2.0.0",
  "streamlit-aggrid>=0.3.4,<1.0.0",
  "plotly>=5.18.0,<6.0.0",

  # Azure / Power BI auth
  "azure-identity>=1.15.0,<2.0.0",
  "msal>=1.26.0,<2.0.0",

  # Formatting
  "tabulate>=0.9.0,<1.0.0",
]

[project.optional-dependencies]
# Extra groups map to pip extras: .[analytics], .[reporting], .[test]
analytics = [
  "pandas>=2.0.0,<3.0.0",
  "numpy>=1.24.0,<2.0.0",
  "scipy>=1.10.0,<2.0.0",
  "statsmodels>=0.14.0,<1.0.0",
  "plotly>=5.18.0,<6.0.0",
  "matplotlib>=3.7.0,<4.0.0",
  "seaborn>=0.12.0,<1.0.0",
  "scikit-learn>=1.3.0,<2.0.0",
]

reporting = [
  "fpdf2>=2.7.0,<3.0.0",
  "python-pptx>=0.6.23,<1.0.0",
]

nlp = [
  "spacy>=3.7.2,<4.0.0",
  "transformers>=4.37.0,<5.0.0",
  "sentence-transformers>=2.3.1,<3.0.0",
  "torch>=2.1.0,<3.0.0",
  "torchvision>=0.16.0,<1.0.0",
  "torchaudio>=2.1.0,<3.0.0",
]

test = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "pytest-cov>=4.1.0,<6.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
asyncio_mode = "strict"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
  ".venv",
  "data",
  "exports",
  "reports",
  "**/__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
